requirements.txt
flask==2.0.1
spacy==3.4.1
python-docx==0.8.11
pdfminer.six==20211012
scikit-learn==1.0.2